refactor to add coinbase
